services/features/pyproject.toml
[project]
name = "features"
version = "0.1.0"
description = "Pipeline 2: Feature engineering from raw Kafka streams to RisingWave"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "confluent-kafka>=2.0",
    "psycopg2-binary>=2.9",
    "structlog>=24.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
